refactor the project name updated readme
