+++ src/main/java/com/glitterlabs/terraformui/controller/DashboardController.java
@@ -1,4 +1,4 @@
-package com.glitterlabs.project.controller;
+package com.glitterlabs.terraformui.controller;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
