+++ src/main/java/com/glitterlabs/terraformui/dao/CloudRepository.java
@@ -1,9 +1,9 @@
-package com.glitterlabs.project.dao;
+package com.glitterlabs.terraformui.dao;
 
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.data.repository.CrudRepository;
 
-import com.glitterlabs.project.model.Cloud;
+import com.glitterlabs.terraformui.model.Cloud;
 
 public interface CloudRepository extends CrudRepository<Cloud, Long> {
 	@Cacheable("clouds")
