+++ src/main/java/com/glitterlabs/terraformui/model/Resource.java
@@ -1,4 +1,4 @@
-package com.glitterlabs.project.model;
+package com.glitterlabs.terraformui.model;
 
 public class Resource {
