+++ src/main/resources/global.properties
@@ -1,1 +1,2 @@
 directory.root=D:\\myProjects\\terraUI\\directory
+terraform.exe.path=D:\\myProjects\\terraUI\\terraform.exe
